Read-back, set_*_data(), and panel edits that survive refresh()
API
- heatmap()/imshow() take std::span<const double> (breaking); values are
  still stored as float.
- Read-back on Axes, Plane2D and Axes3D: title()/xtitle()/ytitle()/ztitle(),
  xlim()/ylim()/zlim() as drawn (via a resolve_limits() shared with layout),
  and <kind>_count()/<kind>_data(i) returning LineData, ScatterData,
  ScatterZData, BarData, HeatmapData, Bar3DData, SurfaceData,
  SurfaceTriData, Scatter3DData, Line3DData.
- set_<kind>_data(i, data) for every read-back kind: replaces an object's
  data in place, validated as plotting it, with nothing changed if it
  throws. Error bars and hint_labels are kept while the point count is
  unchanged. Bar widths are kept unless x (u/v) changes. surface_tri
  topology is taken as given.

Panel edits
- Every panel edit is now journaled and replayed into Axes::Impl /
  Figure::Impl on refresh(), so a program that refreshes in a loop no
  longer wipes them. Data ops replay in order; everything else (titles,
  limits incl. 2D pan/zoom, 3D camera, styles, grid, legend, colorbar,
  ticks, plot/object appearance, 3D box/aspect/planes, suptitle, margins)
  replays as its latest value.
- Stamps keep the program's setter calls authoritative. Title, limit
  (per axis), camera and style-group setters stamp their fields; a panel
  edit records the stamps of the snapshot it was made over and applies
  only if nothing newer was set. FigureEditBox::set_drawn() records the
  stamps of the snapshot each frame's panels are drawn from.
- Index guards: a data op carries its plot's data_stamp, so it drops once
  the data is replaced or re-plotted. Object and plane edits carry the
  cla()/placement stamp, so they can't land on a reused index.
- The panel re-seeds its camera and Limits fields when the program sets
  them.
- Shared templates apply_axes_edit(), apply_style_edits() and
  apply_figure_edits() replace three copies of the apply code.

Tests
- layout_test: read-back, set_*_data (2D/3D), keep_aligned, data-op
  stamps, and title, camera, limit and style journals (1720 checks).
- window_probe --manual animates a 2D and a 3D window with set_*_data()
  and refresh() every frame; window_test's live-refresh demo uses
  set_line_data().

+++ tests/layout_test/heatmap.cpp
@@ -133,7 +133,7 @@ namespace lt {
 
         // --- imshow() is heatmap() at the index extent; indivisible ranges throw.
         {
-            const std::vector<float> data(25, 0.0f);
+            const std::vector<double> data(25, 0.0);
             auto fig = sextant::Figure::create({.width = 400, .height = 300});
             bool threw_degenerate = false, threw_nonfinite = false, imshow_ok = true;
             try { fig->axes()->heatmap(data, 5, 5, {1.0, 1.0}, {0.0, 5.0}); } catch (const std::invalid_argument&) {
